+++ lib/edytuj_z_menu/edytuj_z_menu_widget.dart
@@ -1109,8 +1109,8 @@ class _EdytujZMenuWidgetState extends State<EdytujZMenuWidget> {
                                 ...createProductsRecordData(
                                   name: _model.textController1.text,
                                   description: _model.textController5.text,
-                                  price:
-                                      int.tryParse(_model.textController2.text),
+                                  price: double.tryParse(
+                                      _model.textController2.text),
                                   calories:
                                       int.tryParse(_model.textController3.text),
                                   category: _model.dropDownValue ==
